+++ src/Xrefcheck/System.hs
@@ -5,12 +5,14 @@
 
 module Xrefcheck.System
     ( readingSystem
-    , CanonicalizedGlobPattern (..)
+    , RelGlobPattern (..)
+    , bindGlobPattern
     ) where
 
 import Data.Aeson (FromJSON (..), withText)
 import GHC.IO.Unsafe (unsafePerformIO)
 import System.Directory (canonicalizePath)
+import System.FilePath ((</>))
 import qualified System.FilePath.Glob as Glob
 
 -- | We can quite safely treat surrounding filesystem as frozen,
@@ -18,12 +20,28 @@ import qualified System.FilePath.Glob as Glob
 readingSystem :: IO a -> a
 readingSystem = unsafePerformIO
 
--- | Glob pattern with 'canonicalizePath' applied O_o.
-newtype CanonicalizedGlobPattern = CanonicalizedGlobPattern Glob.Pattern
+-- | Glob pattern relative to repository root.
+newtype RelGlobPattern = RelGlobPattern FilePath
 
-instance FromJSON CanonicalizedGlobPattern where
-    parseJSON = withText "Repo-rooted glob pattern" $ \path -> do
-        let !cpath = readingSystem $ canonicalizePath (toString path)
-        cpat <- Glob.tryCompileWith Glob.compDefault cpath
-                & either fail pure
-        return $ CanonicalizedGlobPattern cpat
+bindGlobPattern :: FilePath -> RelGlobPattern -> Glob.Pattern
+bindGlobPattern root (RelGlobPattern relPat) = readingSystem $ do
+  -- TODO [#26] try to avoid using canonicalization
+  absPat <- canonicalizePath (root </> relPat)
+  case Glob.tryCompileWith globCompileOptions absPat of
+    Left err ->
+      error $ "Glob pattern compilation failed after canonicalization: " <>
+              toText err
+    Right pat ->
+      return pat
+
+instance FromJSON RelGlobPattern where
+    parseJSON = withText "Repo-relative glob pattern" $ \path -> do
+        let spath = toString path
+        -- Checking path is sane
+        _ <- Glob.tryCompileWith globCompileOptions spath
+             & either fail pure
+        return (RelGlobPattern spath)
+
+-- | Glob compilation options we use.
+globCompileOptions :: Glob.CompOptions
+globCompileOptions = Glob.compDefault